+++ Molly/src/App.tsx
@@ -17,6 +17,7 @@ import { ClipsListView } from './views/Clips/ClipsListView';
 import { C4SView } from './views/C4S/C4SView';
 import { BundlesListView } from './views/Bundles/BundlesListView';
 import { GifStudioView } from './views/GifStudio/GifStudioView';
+import { SquishView } from './views/Squish/SquishView';
 import { JobsView } from './views/Jobs/JobsView';
 import { NotesView } from './views/Notes/NotesView';
 import { UserManualView } from './views/Manual/UserManualView';
@@ -134,6 +135,7 @@ export default function App() {
     case 'c4s':       body = <C4SView active={active} />; break;
     case 'bundles':   body = <BundlesListView active={active} />; break;
     case 'gifstudio': body = <GifStudioView />; break;
+    case 'squish':    body = <SquishView />; break;
     case 'jobs':      body = <JobsView />; break;
     case 'notes':     body = <NotesView />; break;
     case 'customers': body = <CustomerListView active={active} />; break;